# Speedup `compress` clause insertion