Elpi's compiler is slow/naive when it comes to grafting clauses.
This PR rephrases the `compress` predicate so that we don't need to insert clauses `:before` the anchor point.

On my laptop on 8.17 we go from
```
real	34m47,869s
user	34m23,509s
sys	0m23,660s
```
to
```
real	26m54,973s
user	26m32,587s
sys	0m22,236s
```