@proux01 2023-06-14
Apparently it works perfectly: https://github.com/math-comp/math-comp/pull/1038